ADD 공식, 유닛옵션

+++ Assets/JKH/00.Scripts/NEW/Unit.cs
@@ -9,7 +9,7 @@ public class Unit : MonoBehaviour
     public bool isSelected;
 
     //Unit's stat
-    public int movePower = 2;
+    public float movePower = 2;
     public int hp = 100;
